Fix some compiler warnings (implicit casts, wrong return datatypes)

+++ include/parserutils/charset/utf8.h
@@ -31,13 +31,13 @@ parserutils_error parserutils_charset_utf8_length(const uint8_t *s, size_t max,
 parserutils_error parserutils_charset_utf8_char_byte_length(const uint8_t *s,
 		size_t *len);
 
-parserutils_error parserutils_charset_utf8_prev(const uint8_t *s, uint32_t off,
-		uint32_t *prevoff);
-parserutils_error parserutils_charset_utf8_next(const uint8_t *s, uint32_t len,
-		uint32_t off, uint32_t *nextoff);
+parserutils_error parserutils_charset_utf8_prev(const uint8_t *s, off_t off,
+		off_t *prevoff);
+parserutils_error parserutils_charset_utf8_next(const uint8_t *s, size_t len,
+		off_t off, off_t *nextoff);
 
 parserutils_error parserutils_charset_utf8_next_paranoid(const uint8_t *s, 
-		uint32_t len, uint32_t off, uint32_t *nextoff);
+		size_t len, off_t off, off_t *nextoff);
 
 #ifdef __cplusplus
 }
